move bitarray serialization to serialization.py
