+++ clkhash/bloomfilter.py
@@ -345,20 +345,3 @@ def stream_bloom_filters(dataset,  # type: Iterable[Sequence[Text]]
                                 keys, hash_properties)
             for s in dataset)
 
-
-def serialize_bitarray(ba):
-    # type: (bitarray) -> str
-    """Serialize a bitarray (bloomfilter)
-
-    """
-    return base64.b64encode(ba.tobytes()).decode('utf8')
-
-# TODO: I think this must already exist in anonlink - maybe that implelemtation should be moved here
-def deserialize_bitarray(ser):
-    # type: (str) -> bitarray
-    """Deserialize a base 64 encoded string to a bitarray (bloomfilter)
-    
-    """
-    ba = bitarray()   # TODO: Is endianess used consistently?
-    ba.frombytes(base64.b64decode(ser.encode(encoding='UTF-8', errors='strict')))
-    return ba
