+++ clkhash/cli.py
@@ -4,7 +4,6 @@ from __future__ import print_function
 import json
 import os
 import shutil
-import time
 
 import click
